+++ Assets/Content/Data/DefaultPrefabObjects.asset
@@ -22,6 +22,8 @@ MonoBehaviour:
   - {fileID: 2930813178971533500, guid: 1c3eed1ea37ecc1448e41270cdc63486, type: 3}
   - {fileID: 1871071288925904272, guid: add4d1bc2d874d94782d6b8c912af5e2, type: 3}
   - {fileID: 7942135432533076774, guid: 84a5596b28f449e49bb34741abf89d30, type: 3}
+  - {fileID: 7942135432533076774, guid: 96f34e849ae8c5640a9343b5ef370179, type: 3}
+  - {fileID: 7942135432533076774, guid: d25b39d1e2379b64cba47629530ada78, type: 3}
   - {fileID: -4994790824565799921, guid: a28267798293970468ab30803632a3f7, type: 3}
   - {fileID: 7823849392942818498, guid: 854451834f026fb4f860bf0e977a104c, type: 3}
   - {fileID: 7564222845035831833, guid: 3b33b5cf16a677a4e85a110679cde32a, type: 3}
@@ -29,7 +31,12 @@ MonoBehaviour:
   - {fileID: -5633820156739102180, guid: 4c966398d6624164c910cd29c1faa980, type: 3}
   - {fileID: -7828078595381397262, guid: 50dd6be959d44844ab1abf8a3f2adb01, type: 3}
   - {fileID: -7828078595381397262, guid: d07b5fba3896ff448855d6e54356c365, type: 3}
+  - {fileID: -2900075557854325373, guid: 6c2559c2017ada740b245951a23c2e1d, type: 3}
   - {fileID: -2900075557854325373, guid: 45d13347efae0f54882e0e56dd770bb1, type: 3}
+  - {fileID: -2900075557854325373, guid: 9682b163cae64f74a93b6e84ce84ae5b, type: 3}
+  - {fileID: -2900075557854325373, guid: abcc60752234e584d8d98c9dec03e7cc, type: 3}
+  - {fileID: -2900075557854325373, guid: b9c224febb4233f4e912745553097652, type: 3}
+  - {fileID: -2900075557854325373, guid: f2493b5478dd8e946ba98861f533362e, type: 3}
   - {fileID: -7727686409064562107, guid: 92d3eb280886ba74cb1e19f4e01669a7, type: 3}
   - {fileID: -7727686409064562107, guid: 2b932aed8c9bdb54da3d5a0274fe9927, type: 3}
   - {fileID: -7727686409064562107, guid: cc48c05d209320541af6b8785c1eec62, type: 3}
@@ -44,6 +51,7 @@ MonoBehaviour:
   - {fileID: 1019513474262773944, guid: 06c946c7c639138488b814e4448b82e4, type: 3}
   - {fileID: 1871071288925904272, guid: b7976a98a25be4543ba45f1d5e26c072, type: 3}
   - {fileID: 3258417366846669140, guid: fa1909e0d40b4cf408ff599df7a2043f, type: 3}
+  - {fileID: 3258417366846669140, guid: 330d85a67a5891145bd8037dde90f37b, type: 3}
   - {fileID: 3258417366846669140, guid: dd3ce5340ed4aba4babae7020f737084, type: 3}
   - {fileID: 5958074810027370513, guid: cbe4989d1925276409163212a1ee01b8, type: 3}
   - {fileID: 564570081198347949, guid: de543fe84ff1e564e9f616e3aa62c122, type: 3}
