+++ Assets/Content/Data/ObserverConditions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19b30e4084b9f3c4e80d6da3fc9b982c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
