+++ Assets/Scripts/SS3D/Core/ActorLocator.cs
@@ -50,7 +50,7 @@ namespace SS3D.Core
             string message = $"Couldn't find actor of id {id} in the scene";
             
             // ReSharper disable once Unity.PerformanceCriticalCodeInvocation
-            Punpun.Panic(typeof(Subsystems), message, Logs.Important);
+            Punpun.Error(typeof(Subsystems), message, Logs.Important);
 
             return null;
         }
