+++ Assets/Scripts/SS3D/Core/ViewLocator.cs
@@ -66,7 +66,7 @@ namespace SS3D.Core
             string message = $"No views of type {typeof(T).Name} found.";
 
             // ReSharper disable once Unity.PerformanceCriticalCodeInvocation
-            Punpun.Panic(typeof(Subsystems), message, Logs.Important);
+            Punpun.Error(typeof(Subsystems), message, Logs.Important);
 
             return null;
         }
