+++ Assets/Scripts/SS3D/Systems/Tile/PlacedObjects/PlacedItemObject.cs
@@ -39,7 +39,7 @@ namespace SS3D.Systems.Tile
             if (InstanceFinder.ServerManager != null && placedObject.GetComponent<NetworkObject>() != null)
             {
                 if (placedObject.GetComponent<NetworkObject>() == null)
-                    Punpun.Yell(Subsystems.Get<TileSubsystem>(), $"{placedObject.NameString} does not have a Network Component and will not be spawned");
+                    Punpun.Warning(Subsystems.Get<TileSubsystem>(), $"{placedObject.NameString} does not have a Network Component and will not be spawned");
                 else
                     InstanceFinder.ServerManager.Spawn(placedGameObject);
             }
