+++ Assets/Scripts/SS3D/Systems/Tile/TileMapCreator/TileMapCreator.cs
@@ -9,6 +9,7 @@ using FishNet.Object;
 using FishNet.Connection;
 using SS3D.Logging;
 using SS3D.Core.Behaviours;
+using SS3D.Systems.Inputs;
 using SS3D.Systems.Tile.TileMapCreator;
 using UnityEngine.InputSystem;
 
@@ -57,7 +58,7 @@ namespace SS3D.Systems.Tile.UI
         private void Start()
         {
             ShowUI(false);
-            _controls = SystemLocator.Get<InputSystem>().Inputs.TileCreator;
+            _controls = Subsystems.Get<InputSubsystem>().Inputs.TileCreator;
             _controls.ToggleMenu.Enable();
             _controls.ToggleMenu.performed += HandleToggleMenu;
             _controls.Replace.performed += HandleReplace;
@@ -383,14 +384,14 @@ namespace SS3D.Systems.Tile.UI
         public void OnPointerEnter(PointerEventData eventData)
         {
             _mouseOverUI = true;
-            SystemLocator.Get<InputSystem>().Inputs.Camera.Zoom.Disable();
+            Subsystems.Get<InputSubsystem>().Inputs.Camera.Zoom.Disable();
         }
 
         [ServerOrClient]
         public void OnPointerExit(PointerEventData eventData)
         {
             _mouseOverUI = false;
-            SystemLocator.Get<InputSystem>().Inputs.Camera.Zoom.Enable();
+            Subsystems.Get<InputSubsystem>().Inputs.Camera.Zoom.Enable();
         }
     }
 }
